feat: init frames
