+++ README.md
@@ -25,3 +25,9 @@ yarn test
 yarn install
 yarn start:dev
 ```
+
+Development
+```
+ngrok http 3000
+yarn next:frames
+```
