+++ packages/nextjs/app/(site)/blockexplorer/_components/AddressComponent.tsx
