+++ packages/nextjs/app/(site)/blockexplorer/_components/SearchBar.tsx
