+++ packages/nextjs/app/(site)/blockexplorer/address/[address]/page.tsx
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
 import { hardhat } from "viem/chains";
-import { AddressComponent } from "~~/app/blockexplorer/_components/AddressComponent";
+import { AddressComponent } from "~~/app/(site)/blockexplorer/_components/AddressComponent";
 import deployedContracts from "~~/contracts/deployedContracts";
 import { GenericContractsDeclaration } from "~~/src/utils/scaffold-eth/contract";
