+++ packages/nextjs/app/(site)/debug/_components/contract/ContractReadMethods.tsx
@@ -1,24 +1,19 @@
 import { Abi, AbiFunction } from "abitype";
-import { WriteOnlyFunctionForm } from "~~/app/debug/_components/contract";
+import { ReadOnlyFunctionForm } from "~~/app/(site)/debug/_components/contract";
 import { Contract, ContractName, GenericContract, InheritedFunctions } from "~~/src/utils/scaffold-eth/contract";
 
-export const ContractWriteMethods = ({
-  onChange,
-  deployedContractData,
-}: {
-  onChange: () => void;
-  deployedContractData: Contract<ContractName>;
-}) => {
+export const ContractReadMethods = ({ deployedContractData }: { deployedContractData: Contract<ContractName> }) => {
   if (!deployedContractData) {
     return null;
   }
 
   const functionsToDisplay = (
-    (deployedContractData.abi as Abi).filter(part => part.type === "function") as AbiFunction[]
+    ((deployedContractData.abi || []) as Abi).filter(part => part.type === "function") as AbiFunction[]
   )
     .filter(fn => {
-      const isWriteableFunction = fn.stateMutability !== "view" && fn.stateMutability !== "pure";
-      return isWriteableFunction;
+      const isQueryableWithParams =
+        (fn.stateMutability === "view" || fn.stateMutability === "pure") && fn.inputs.length > 0;
+      return isQueryableWithParams;
     })
     .map(fn => {
       return {
@@ -29,18 +24,17 @@ export const ContractWriteMethods = ({
     .sort((a, b) => (b.inheritedFrom ? b.inheritedFrom.localeCompare(a.inheritedFrom) : 1));
 
   if (!functionsToDisplay.length) {
-    return <>No write methods</>;
+    return <>No read methods</>;
   }
 
   return (
     <>
-      {functionsToDisplay.map(({ fn, inheritedFrom }, idx) => (
-        <WriteOnlyFunctionForm
+      {functionsToDisplay.map(({ fn, inheritedFrom }) => (
+        <ReadOnlyFunctionForm
           abi={deployedContractData.abi as Abi}
-          key={`${fn.name}-${idx}}`}
-          abiFunction={fn}
-          onChange={onChange}
           contractAddress={deployedContractData.address}
+          abiFunction={fn}
+          key={fn.name}
           inheritedFrom={inheritedFrom}
         />
       ))}
