+++ packages/nextjs/app/(site)/debug/_components/contract/ContractVariables.tsx
