+++ packages/nextjs/app/(site)/debug/_components/contract/DisplayVariable.tsx
