+++ packages/nextjs/app/(site)/debug/_components/contract/InheritanceTooltip.tsx
