+++ packages/nextjs/app/(site)/debug/_components/contract/utilsDisplay.tsx
