+++ packages/nextjs/app/frames/lottery/create/route.tsx
@@ -1,0 +1,7 @@
+import { frames } from "~~/app/frames/frames";
+
+export const POST = frames(async () => {
+  return {
+    image: <div>You created a lottery</div>,
+  };
+});
