+++ packages/nextjs/package.json
@@ -7,6 +7,7 @@
     "check-types": "tsc --noEmit --incremental",
     "dev": "next dev",
     "format": "prettier --write . '!(node_modules|.next|contracts)/**/*'",
+    "frames": "frames",
     "lint": "next lint",
     "start": "next start",
     "start:dev": "next dev",
@@ -21,6 +22,7 @@
     "@uniswap/v2-sdk": "~3.0.1",
     "blo": "~1.0.1",
     "daisyui": "4.5.0",
+    "frames.js": "^0.15.0",
     "next": "~14.0.4",
     "next-themes": "~0.2.1",
     "nprogress": "~0.2.0",
@@ -36,6 +38,7 @@
     "zustand": "~4.1.2"
   },
   "devDependencies": {
+    "@frames.js/debugger": "^0.2.3",
     "@trivago/prettier-plugin-sort-imports": "~4.1.1",
     "@types/node": "^17.0.45",
     "@types/nprogress": "^0",
